refactor: combine api/client with base.py
 - this makes the Pathy type accessible where it otherwise would not be for TypeVars.
